refact(indexing-process): improves to automatic relate indexing process to pagination
